Refactor task image extraction
